+++ packages/core/src/config/config.ts
@@ -1,168 +1,19 @@
-import type { Abi, AbiEvent, FormatAbiItem } from "abitype";
-import type { GetEventArgs, Transport } from "viem";
+import type { Prettify } from "@/types/utils.js";
+import type { Abi } from "abitype";
+import { type Transport } from "viem";
+import type { GetAddress } from "./address.js";
+import type { GetEventFilter } from "./eventFilter.js";
 
-export type FilterAbiEvents<T extends Abi | readonly unknown[]> =
-  T extends readonly [infer First, ...infer Rest extends Abi]
-    ? First extends AbiEvent
-      ? readonly [First, ...FilterAbiEvents<Rest>]
-      : FilterAbiEvents<Rest>
-    : [];
-
-/**
- * Remove TElement from TArr.
- */
-type FilterElement<
-  TElement,
-  TArr extends readonly unknown[],
-> = TArr extends readonly [infer First, ...infer Rest]
-  ? TElement extends First
-    ? FilterElement<TElement, Rest>
-    : readonly [First, ...FilterElement<TElement, Rest>]
-  : [];
-
-/**
- * Return an array of safe event names that handle event overridding.
- */
-export type SafeEventNames<
-  TAbi extends readonly AbiEvent[],
-  TArr extends readonly AbiEvent[] = TAbi,
-> = TAbi extends readonly [
-  infer First extends AbiEvent,
-  ...infer Rest extends readonly AbiEvent[],
-]
-  ? First["name"] extends FilterElement<First, TArr>[number]["name"]
-    ? // Overriding occurs, use full name
-      FormatAbiItem<First> extends `event ${infer LongEvent extends string}`
-      ? readonly [LongEvent, ...SafeEventNames<Rest, TArr>]
-      : never
-    : // Short name
-      readonly [First["name"], ...SafeEventNames<Rest, TArr>]
-  : [];
-
-/**
- * Recover the element from {@link TAbi} at the index where {@link TSafeName} is equal to {@link TSafeNames}[index].
- */
-export type RecoverAbiEvent<
-  TAbi extends readonly AbiEvent[],
-  TSafeName extends string,
-  TSafeNames extends readonly string[] = SafeEventNames<TAbi>,
-> = TAbi extends readonly [
-  infer FirstAbi,
-  ...infer RestAbi extends readonly AbiEvent[],
-]
-  ? TSafeNames extends readonly [
-      infer FirstName,
-      ...infer RestName extends readonly string[],
-    ]
-    ? FirstName extends TSafeName
-      ? FirstAbi
-      : RecoverAbiEvent<RestAbi, TSafeName, RestName>
-    : never
-  : never;
-
-/** Required fields for a contract. */
-export type ContractRequired<
-  TNetworkNames extends string,
-  TAbi extends Abi | readonly unknown[],
-  TEventName extends string,
-  TFactoryEvent extends AbiEvent | undefined,
-> = {
-  /** Contract application byte interface. */
-  abi: TAbi;
-  /**
-   * Network that this contract is deployed to. Must match a network name in `networks`.
-   * Any filter information overrides the values in the higher level "contracts" property.
-   * Factories cannot override an address and vice versa.
-   */
-  network:
-    | Partial<
-        Record<
-          TNetworkNames,
-          Partial<ContractFilter<TAbi, TEventName, TFactoryEvent>>
-        >
-      >
-    | TNetworkNames;
-};
-
-/** Fields for a contract used to filter down which events indexed. */
-export type ContractFilter<
-  TAbi extends Abi | readonly unknown[],
-  TEventName extends string,
-  TFactoryEvent extends AbiEvent | undefined,
-> = (
-  | {
-      address: `0x${string}` | readonly `0x${string}`[];
-      factory?: never;
-    }
-  | {
-      address?: never;
-      /** Factory contract configuration. */
-      factory: {
-        /** Address of the factory contract that creates this contract. */
-        address: `0x${string}`;
-        /** ABI event that announces the creation of a new instance of this contract. */
-        event: AbiEvent;
-        /** Name of the factory event parameter that contains the new child contract address. */
-        p: TFactoryEvent;
-        parameter: TFactoryEvent extends AbiEvent
-          ? TFactoryEvent["inputs"][number]["name"]
-          : string;
-      };
-    }
-  | {}
-) & {
-  /** Block number at which to start indexing events (inclusive). Default: `0`. */
+export type BlockConfig = {
+  /** Block number at which to start indexing events (inclusive). If `undefined`, events will be processed from block 0. Default: `undefined`. */
   startBlock?: number;
   /** Block number at which to stop indexing events (inclusive). If `undefined`, events will be processed in real-time. Default: `undefined`. */
   endBlock?: number;
   /** Maximum block range to use when calling `eth_getLogs`. Default: `10_000`. */
   maxBlockRange?: number;
-
-  filter?: Abi extends TAbi
-    ?
-        | { event: readonly string[]; args?: never }
-        | { event: string; args?: GetEventArgs<Abi, string> }
-    : TAbi extends Abi
-      ?
-          | {
-              event: readonly SafeEventNames<FilterAbiEvents<TAbi>>[number][];
-            }
-          | {
-              event: SafeEventNames<FilterAbiEvents<TAbi>>[number];
-              args?: GetEventArgs<
-                Abi,
-                string,
-                {
-                  EnableUnion: true;
-                  IndexedOnly: true;
-                  Required: false;
-                },
-                RecoverAbiEvent<
-                  FilterAbiEvents<TAbi>,
-                  TEventName,
-                  SafeEventNames<FilterAbiEvents<TAbi>>
-                > extends infer _abiEvent extends AbiEvent
-                  ? _abiEvent
-                  : AbiEvent
-              > extends infer e
-                ? e extends readonly []
-                  ? never
-                  : e
-                : never;
-            }
-      : never;
 };
 
-/** Contract in Ponder config. */
-export type Contract<
-  TNetworkNames extends string,
-  TAbi extends Abi | readonly unknown[],
-  TEventName extends string,
-  TFactoryEvent extends AbiEvent | undefined,
-> = ContractRequired<TNetworkNames, TAbi, TEventName, TFactoryEvent> &
-  ContractFilter<TAbi, TEventName, TFactoryEvent>;
-
-type Database =
+export type DatabaseConfig =
   | {
       kind: "sqlite";
       /** Path to SQLite database file. Default: `"./.ponder/cache.db"`. */
@@ -174,8 +25,12 @@ type Database =
       connectionString?: string;
     };
 
-/** Network in Ponder config. */
-export type Network = {
+export type OptionConfig = {
+  /** Maximum number of seconds to wait for event processing to be complete before responding as healthy. If event processing exceeds this duration, the API may serve incomplete data. Default: `240` (4 minutes). */
+  maxHealthcheckDuration?: number;
+};
+
+export type NetworkConfig = {
   /** Chain ID of the network. */
   chainId: number;
   /** A viem `http`, `webSocket`, or `fallback` [Transport](https://viem.sh/docs/clients/transports/http.html).
@@ -192,76 +47,101 @@ export type Network = {
    * }
    * ```
    */
-  transport: Transport | unknown;
+  transport: Transport;
   /** Polling frequency (in ms). Default: `1_000`. */
   pollingInterval?: number;
   /** Maximum concurrency of tasks during the historical sync. Default: `20`. */
   maxHistoricalTaskConcurrency?: number;
 };
 
-type Option = {
-  /** Maximum number of seconds to wait for event processing to be complete before responding as healthy. If event processing exceeds this duration, the API may serve incomplete data. Default: `240` (4 minutes). */
-  maxHealthcheckDuration?: number;
+type ContractRequired<
+  networks extends { [name: string]: unknown },
+  abi extends Abi | readonly unknown[],
+  ///
+  contractNetwork extends keyof networks = keyof networks,
+  allNetworkNames = keyof networks,
+> = {
+  /** Contract application byte interface. */
+  abi: abi;
+  /**
+   * Network that this contract is deployed to. Must match a network name in `networks`.
+   * Any filter information overrides the values in the higher level "contracts" property.
+   * Factories cannot override an address and vice versa.
+   */
+  network:
+    | allNetworkNames
+    | (contractNetwork extends allNetworkNames ? contractNetwork : never);
 };
 
-export type Config = {
-  /** Database to use for storing raw & indexed data. Default: `"postgres"` if `DATABASE_URL` env var is present, otherwise `"sqlite"`. */
-  database?: Database;
-  networks: Record<string, Network>;
-  /** List of contracts to sync & index events from. Contracts defined here will be present in `context.contracts`. */
-  contracts: Record<string, Contract<string, Abi, string, AbiEvent>>;
-  /** Configuration for Ponder internals. */
-  options?: Option;
+type GetContract<
+  networks extends { [name: string]: unknown },
+  contract,
+> = contract extends {
+  abi: infer abi extends Abi;
+}
+  ? // 1. Contract has a valid abi
+    contract extends { network: infer contractNetwork extends keyof networks }
+    ? // 1.a Contract has a valid abi and network
+      Prettify<
+        ContractRequired<networks, abi, contractNetwork> &
+          GetAddress<Omit<contract, "network" | "abi" | "filter">> &
+          GetEventFilter<
+            abi,
+            Omit<contract, "network" | "address" | "factory" | "abi">
+          >
+      >
+    : // 1.b Contract has valid abi and invalid network
+      Prettify<ContractRequired<networks, abi>> &
+        GetAddress<Omit<contract, "network" | "abi" | "filter">> &
+        GetEventFilter<
+          abi,
+          Omit<contract, "network" | "address" | "factory" | "abi">
+        >
+  : // 2. Contract has an invalid abi
+    contract extends { network: infer contractNetwork extends keyof networks }
+    ? // 2.a Contract has an invalid abi and a valid network
+      Prettify<
+        ContractRequired<networks, Abi, contractNetwork> &
+          GetAddress<Omit<contract, "network" | "abi" | "filter">> &
+          GetEventFilter<
+            Abi,
+            Omit<contract, "network" | "address" | "factory" | "abi">
+          >
+      >
+    : // 2.b Contract has an invalid abi and an invalid network
+      Prettify<
+        ContractRequired<networks, Abi> &
+          GetAddress<Omit<contract, "network" | "abi" | "filter">> &
+          GetEventFilter<
+            Abi,
+            Omit<contract, "network" | "address" | "factory" | "abi">
+          >
+      >;
+
+type ContractsConfig<
+  networks extends { [name: string]: unknown },
+  contracts,
+> = {} extends contracts // contracts empty, return empty
+  ? {}
+  : {
+      [name in keyof contracts]: GetContract<networks, contracts[name]>;
+    };
+
+type NetworksConfig<networks extends { [name: string]: unknown }> = {
+  [networkName in keyof networks]: NetworkConfig;
 };
 
-/**
- * Create a Ponder config.
- *
- * - Docs: https://ponder.sh/docs/api-reference/config
- *
- * @example
- * export default createConfig({
- *   networks: {
- *     mainnet: { chainId: 1, transport: http(process.env.PONDER_RPC_URL_1) },
- *   },
- *   contracts: {
- *     ERC20: {
- *       abi: ERC20Abi,
- *       network: "mainnet",
- *       address: "0x123...",
- *       startBlock: 1500,
- *     },
- *   },
- * })
- */
 export const createConfig = <
-  const TNetworks extends Record<string, Network>,
-  const TContracts extends {
-    [ContractName in keyof TContracts]: Contract<
-      keyof TNetworks & string,
-      TContracts[ContractName]["abi"] & (Abi | readonly unknown[]),
-      TContracts[ContractName]["filter"] extends {
-        event: infer _event extends string;
-      }
-        ? _event
-        : string,
-      TContracts[ContractName] extends {
-        factory: {
-          event: infer _event extends AbiEvent;
-        };
-      }
-        ? _event
-        : undefined
-    >;
-  },
+  const networks extends { [name: string]: unknown },
+  const contracts extends { [name: string]: unknown },
 >(config: {
-  database?: Database;
-  networks: TNetworks;
-  contracts: TContracts;
-  options?: Option;
-}): {
-  database?: Database;
-  networks: TNetworks;
-  contracts: TContracts;
-  options?: Option;
-} => config;
+  // TODO: add jsdoc to these properties.
+  networks: NetworksConfig<networks>;
+  contracts: ContractsConfig<networks, contracts>;
+  database?: DatabaseConfig;
+  options?: OptionConfig;
+}) => {
+  return config;
+};
+
+export type Config = Parameters<typeof createConfig>[0];
